Last couple of portability fixes for Solaris

+++ hdf/test/rig.c
@@ -1054,7 +1054,7 @@ static uint8 FAR jpeg_24bit_j75[JPEGY][JPEGX][3]={
 static uint8 FAR jpeg_24bit_temp[JPEGY][JPEGX][3];
 
 
-#define ABS(x)  ((x)<0 ? (-x) : x)
+#define ABS(x)  ((int)(x)<0 ? (-x) : x)
 
 #ifdef PROTOTYPE
 intn fuzzy_memcmp(const void *s1,const void *s2,int32 len,intn fuzz_factor)
